refactor: template section into common template gallery carousel
