+++ apps/journeys-admin/src/components/TemplateGallery/TemplateGalleryCarousel/index.ts
@@ -1,0 +1,1 @@
+export { TemplateGalleryCarousel } from './TemplateGalleryCarousel'
