--- apps/journeys-admin/src/components/TemplateSections/TemplateSection/index.ts
@@ -1,1 +1,0 @@
-export { TemplateSection } from './TemplateSection'
