+++ apps/journeys-admin/src/components/TemplateSections/TemplateSections.tsx
@@ -1,15 +1,17 @@
 import Paper from '@mui/material/Paper'
 import Stack from '@mui/material/Stack'
+import { useTheme } from '@mui/material/styles'
 import Typography from '@mui/material/Typography'
 import map from 'lodash/map'
 import take from 'lodash/take'
 import { ReactElement, useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import { SwiperOptions } from 'swiper'
 
 import { GetJourneys_journeys as Journey } from '../../../__generated__/GetJourneys'
 import { useJourneysQuery } from '../../libs/useJourneysQuery'
-
-import { TemplateSection } from './TemplateSection'
+import { TemplateGalleryCarousel } from '../TemplateGallery/TemplateGalleryCarousel'
+import { TemplateGalleryCard } from '../TemplateGalleryCard'
 
 interface Contents {
   [key: string]: { category: string; journeys: Journey[] }
@@ -25,6 +27,7 @@ export function TemplateSections({
   languageId
 }: TemplateSectionsProps): ReactElement {
   const { t } = useTranslation('apps-journeys-admin')
+  const { breakpoints } = useTheme()
   const [contents, setContents] = useState<Contents>({})
   const [collection, setCollection] = useState<Journey[]>([])
 
@@ -64,12 +67,47 @@ export function TemplateSections({
     }
   })
 
+  const swiperBreakpoints: SwiperOptions['breakpoints'] = {
+    [breakpoints.values.xs]: {
+      slidesPerGroup: 2,
+      slidesPerView: 2,
+      spaceBetween: 20
+    },
+    [breakpoints.values.sm]: {
+      slidesPerGroup: 3,
+      slidesPerView: 3,
+      spaceBetween: 20
+    },
+    [breakpoints.values.md]: {
+      slidesPerGroup: 4,
+      slidesPerView: 4,
+      spaceBetween: 20
+    },
+    [breakpoints.values.lg]: {
+      slidesPerGroup: 5,
+      slidesPerView: 5,
+      spaceBetween: 48
+    },
+    [breakpoints.values.xl]: {
+      slidesPerGroup: 6,
+      slidesPerView: 6,
+      spaceBetween: 48
+    },
+    [breakpoints.values.xxl]: {
+      slidesPerGroup: 7,
+      slidesPerView: 7,
+      spaceBetween: 48
+    }
+  }
+
   return (
     <Stack spacing={8}>
       {(loading || (collection != null && collection.length > 0)) && (
-        <TemplateSection
-          category={tagIds == null ? t('Featured & New') : t('Most Relevant')}
-          journeys={collection}
+        <TemplateGalleryCarousel
+          heading={tagIds == null ? t('Featured & New') : t('Most Relevant')}
+          items={collection}
+          renderItem={(itemProps) => <TemplateGalleryCard {...itemProps} />}
+          breakpoints={swiperBreakpoints}
           loading={loading}
         />
       )}
@@ -98,7 +136,12 @@ export function TemplateSections({
         ({ category, journeys }, key) =>
           ((tagIds == null && journeys.length >= 5) ||
             tagIds?.includes(key) === true) && (
-            <TemplateSection category={category} journeys={journeys} />
+            <TemplateGalleryCarousel
+              heading={category}
+              items={journeys}
+              renderItem={(itemProps) => <TemplateGalleryCard {...itemProps} />}
+              breakpoints={swiperBreakpoints}
+            />
           )
       )}
     </Stack>
